Remove proposal for wallet structure from BIP32.
It remained unused and it is not compatible to BIP43 and the BIPs based on it. If it turns out worthy to keep as a standard, I suggest reviving it under a new BIP number.

As the removed part of the BIP was optional, this change of the standard does not change the compliance state of any BIP32 implementation.