Organize bips by status
Separate BIPS into sections: [Accepted ones, Draft ones, Rejects]. BIPs are sorted into sections and then by number. This makes it still very easy to find BIPs by number but it makes it crystal clear what their status is.

The idea here is to further guide readers towards the concept that a BIP can be a draft or rejected.