Public parent to child parent
I'm really not sure as I'm starting in this field. 
But the last operation to get the public child key from parent public key is a point operation on an integer thus : point(parse<sub>256</sub>(I<sub>L</sub>) + K<sub>par</sub>) and not a point plus adding an interger.