Private key derivation operation order fix
Added additional parentesis for clear understanding of the equation on this row. I am doing this proposition because I was stuck on this part for quite some time untill I understood that I furst need to add and then do the modular devision. I think that this way, the order of operations is more clear.