Update bip-0002.mediawiki
Change workflow:

14.0 How to add a new language

Whoo boy, this is the big one. Please follow all directions exactly.
14.1 _config.yml file

Navigate to the root folder of the whole website and find the file labeled _config.yml. Open it and find the line that says languages:. Add your two letter language code (Google it if you don't know it) in between the brackets after the others already present. You will need to put a comma after the previous last one.

Example:

languages: ["en", "es", "NEW LANG HERE"]

Save and exit the file.
14.2 _data folder

Navigate to the _data/lang folder and copy the template folder. Paste it into the same folder and the copy renamed to the two letter language code of the language you will be translated to. The template folder itself should still be there. It should not be renamed. There should be a new folder in addition to the ones that were already there.

Translate the content of the files. Do not touch anything labeled url, and in the roadmap.yml ONLY translate the name: content.
14.3 _i18n folder

Navigate to the _i18n folder and duplicate the template.yml file. Rename the duplicate to the two letter language code of your language with a .yml at the end. Repeat this process with the template folder as well. The template folder and yml file themselves should still be there. They should not be renamed. There should be a new folder and yml file in addition to the ones that were already there.

Enter the .yml file and translate everything there.
14.3 Translate

If you will be translating the content, please refer to section 13.0 in this README for instructions on translating pages.
14.4 Build/Test

Build your website using jekyll serve. If the build is successful, navigate to any page on the site and check to make sure that your language is appearing in the dropdown for languages on both mobile and desktop.
14.5 Submit a Pull Request

You're all done. Submit a PR and wait for it to be reviewed and merged. Be sure to make any changes if requested.