BIP0173: Include bech32 prefix for regtest mode
A mismatch between implementation occurred on the bech32 prefix for
regtest mode.  Bitcoin Core decided to use the `bcrt` prefix in order
for it to be distinguishable from the testnet prefix.  btcd on the other
hand uses the same prefix as testnet because of lack of a separate
prefix specified in this BIP.

Core: https://github.com/bitcoin/bitcoin/issues/12314
btcd: https://github.com/btcsuite/btcd/pull/1150