BIP8: replace FAILING with MUST_SIGNAL
This removes the FAILING state and replaces compulsory signalling during
LOCKED_IN with compulsory signalling during a new MUST_SIGNAL phase during
the last retarget period prior to the timeout height.  Also reduces the
amount of blocks that are required to signal to match the threshold.

This ensures that if a deployment occurs using bip8 with
lockinontimeout=false and timeoutheight=N, that a later deployment using
bip8 with lockinontimeout=true and timeoutheight=K, where K<N that any
chain where LOCKED_IN is reached prior to height K, will be accepted as
valid by nodes using either set of deployment parameters.

It also ensures that the soft-fork's changed rules are only enforced
on chain a retarget period after signalling indicates enforcement is
expected (which was not previously the case if the FAILING to ACTIVE
transition took place).