bip-340: recreate batch verify speedup graph w/ latest libsecp256k1
Since creating that graph there have been significant improvements in
libsecp256k1 to the speed of certain operations, such as making use of the
endomorphism by default and the safegcd algorithm. This new graph shows the
batch verification speedup in light of these improvements.

There are also some minor improvements to the graph itself. In particular,
the (1,1) datapoint was removed because it is not part of the estimation for the
fitted line.

This commit also contains the Makefile, shell script and gnuplot file that were
used to create the graph.