Remove transparent background from figure.
Before this change, the figure presented black text on transparent
background, which might be unconvenient when using a browser able to
pass a dark theme preference to some environments where this document
is published, currently notably GitHub. A white background could help
a better visualization compromise. White background on the figure is
the single purpose of this revision.

This PNG was compiled using:

    dot -Tpng states.gv -o states.png