BIP-0342: tighten up OP_CEHCKSIGVERIFY tapscript execution description
Upon my initial reading of this section of the BIP, I thought that this
mean that `OP_CEHCKSIGVERIFY` _didn't_ modify the stack by popping up
the element pushed on by `OP_CEHCKSIG`. Instead it's exactly the same,
as pre tapscript behavior.

In this commit, we attempt to make this clearer by removing the "without
any further changes to the stack" text in the BIP. This is what led me
to believe that the op code no longer modified the stack.