BIP370: introduce PSBT_IN_PREVIOUS_OUTPOINT
There is no single use I am aware of for `PSBT_IN_PREVIOUS_TXID` and `PSBT_IN_OUTPUT_INDEX` to be used separately. Thus, I propose to keep all data about the spent output in the input map as a single data structure/key. This helps implementation and simplifies error handling: the situation where just `txid` or `vout` with the second part of the data missed for the spent outpoint are present now will be impossible.