BIP341: Define c[0] = leaf version + pubkey parity
Add explicit definition of _c[0]_ as the _leaf version_ and the public
key _parity bit_.

Prior to this change that interesting fact is only implicit in the code
of `taproot_tweak_pubkey` and `taproot_sign_script`.  Making it explicit
will help reader understanding of the control block and the script
validation rules.