BIP-341: Explain the 64-byte signature format
The signature would be expected to be 65-bytes: the 64-byte Schnorr
signature plus the appended sighash byte.  But the new sighash value
SIGHASH_DEFAULT, encoded as 0x00, allows the space optimization of
omitting the sighash byte, allowing for a 64-byte signature.

This is now explained explicitly in the text.

Address PR comments

- remove ref to `bitcoin.it` wiki - may change or disappear in future (just don't
  refer to a definition)
- Slight wording change to sentence
- Make markup consistent with rest of file