Create japanese_seasonal.txt
The current list of Japanese words is so bad that we have created a word list to describe the beautiful four seasons of Japan. These words are from the Saijiki. The Saijiki is said to have originated in the "Nihon Saijiki" (Japanese chronicle of the seasons) by Mashiken Kaihara in 1688 (Jokyo 5), published by Kyoto Nisshindo. The "Nihon Saijiki" describes seasonal customs, events, flora and fauna, with illustrations of events and daily life in the early Edo period.