Update bip-0039.mediawiki
Just a minor correction on the spec.

The current spec says 'HMAC-SHA512' but this is actually incorrect and possibly misleading. HMAC is actually not used at all in the PKDF part of the spec.

Instead, it should just say 'SHA-512', which is correct.

I had to do a double-take when reading because BIP0032 does use an HMAC-SHA512 method, however I have confirmed with multiple implementations of this spec and everyone uses SHA-512, not HMAC-SHA512.