Require BIPs be "beyond the ideation phase" before being merged
As the BIP process restarts, one general risk in the BIPs is that
they become (even more of) a dumping ground for half-baked ideas
which even the BIP champion does not work on after the BIP is
assigned. While simply never merging new BIPs addresses this issue,
this isn't particularly sustainable, and we should instead have
some concrete criteria here.

BIPs generally fall into a few categories with respect to when they
should be considered ready for merge:
 * they're not particularly controversial (and not about consensus)
   and move forward towards adoption in several ecosystem projects
   relatively quickly or,
 * they generate lots of discussion, lots of feedback, the champion
   actively pushes the BIP forward and makes sustained arguments for
   adoption over some extended time period (this generally applies
   to consensus changes).

As long as the BIP champion is relatively actively pushing for
adoption (i.e. active on the BIP outside of the BIP text itself)
and actively addressing feedback on the BIP, the BIP should be
merged, however if the BIP hasn't seen any adoption anywhere
(outside of maybe one project the author wrote) and the BIP
champion is off doing other things, I'd argue it clearly should
simply be closed.

This commit adds some trivial but flexible text to that effect,
requiring roughly the above criteria for merge.