README: remove BIP 63 from the README index
In this commit, I propose we remove BIP 63 from the README index. The
BIP number was assigned in [this
PR](https://github.com/bitcoin/bips/pull/14), but no document was ever
committed to the repo (?).

Alternatively, based on BIP 2, we could just move it to the Rejected
status, as the PR above was originally made more than 10 years ago:

> BIPs should be changed from Draft or Proposed status, to Rejected
> status, upon request by any person, if they have not made progress in
> three years.